admin: Migrate Table Excel export stories
Move three stories that demonstrate the export hooks against the products
query:

- TableExportAllPages — exports every page using useExportPagedTableQuery
- TableExportWithLimit — exports a fixed number of rows via useExportTableQuery
- TableExportWithLimitFilter — same as above but with an active search filter
